requirements.txt
pandas
numpy==1.26.4
pandas_ta
pykrx
ccxt
matplotlib
parmap
python-telegram-bot==13.4.1
websocket
rel
gdown
psutil
pybit
pycryptodome
Crypto
pykrx
black
mojito2
MetaTrader5
mypy
black